+++ src/lsp/cobol_preproc/src_format.mli
@@ -22,6 +22,7 @@ and fixed_paging_params =
   {
     cut_at_col: int;
     alphanum_padding: char option;
+    enforceable_area_a: bool;
   }
 
 (* Actual format and indicator positioning *)
@@ -58,3 +59,4 @@ val guess_from: contents_prefix: string -> Cobol_config.source_format
 
 val comment_entry_termination: _ source_format -> comment_entry_termination
 val first_area_b_column: _ source_format -> int option
+val enforceable_area_a: _ source_format -> bool